refactor: Consolidate CraneModelService into services module
This commit fixes a `ModuleNotFoundError` by moving the `CraneModelService`
class from its own file (`crane_model_service.py`) into the main
`server/domain/services.py` module.

This aligns the new service with the existing project architecture where
all services are defined and instantiated in a single file, resolving
the import error in the `crane_models` API router.
